packages/socd/pyproject.toml
[project]
name = "socd"
version = "0.1.0"
description = "Library for simulating self-oscillating Class-D amplifiers"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "gmpy2>=2.2.1",
    "numba>=0.60",
    "setuptools>=80.9.0",
    "symengine>=0.14.1",
    "sympy>=1.14.0",
]
[project.optional-dependencies]
cuda = [
    "cupy-cuda12x[all]>=13.4.1",
    "numba-cuda>=0.11.0",
]
[build-system]
requires = ["uv_build>=0.7.12,<0.8.0"]
build-backend = "uv_build"
